// fe/src/lib/api/generated/model/counselingEncounterResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentLinkResponse } from './appointmentLinkResponse';
import type { CounselingEntryMode } from './counselingEntryMode';
import type { DeliveryMode } from './deliveryMode';
import type { IdentitySummaryResponse } from './identitySummaryResponse';
import type { ServiceSummaryResponse } from './serviceSummaryResponse';

export interface CounselingEncounterResponse {
  appointment: AppointmentLinkResponse | null;
  counselor: IdentitySummaryResponse;
  created_at: string;
  delivery_mode: DeliveryMode;
  ended_at: string;
  entry_mode: CounselingEntryMode;
  id: string;
  service: ServiceSummaryResponse;
  started_at: string;
  student: IdentitySummaryResponse;
  updated_at: string;
}